requirements.txt
django
django-rest-framework
python-dotenv
mysqlclient
django-mysql